refactor command line unit tests
